ReLu activation application
these lines apply the ReLU activation function to the output of the fully connected layers
self.fc3 and self.fc4. ReLU introduces the network to non-linearity
and helps the model learn complex patterns and relationships.
The x variable is updated with the outputs after each ReLU
activation and will be used

Links I use:
https://machinelearningmastery.com/rectified-linear-activation-function-for-deep-learning-neural-networks/
https://pytorch.org/tutorials/beginner/blitz/neural_networks_tutorial.html